+++ src/components/RadioGroup/index.ts
@@ -1,1 +1,2 @@
 export { default as RadioGroup } from './RadioGroup.vue'
+export { default as RadioItem } from './RadioItem.vue'
