--- src/components/RadioItem/index.ts
@@ -1,1 +1,0 @@
-export { default as RadioItem } from './RadioItem.vue'
